# [Fizz] Pipeable Stream Perf