--- a/fixtures/fizz/src/BigComponent.js
+++ b/fixtures/fizz/src/BigComponent.js
@@ -1,0 +1,181 @@
+export default function BigComponent() {
+  return (
+    <article>
+      <section>
+        <h2>Description</h2>
+        <p>
+          This page has repeating sections purposefully to create very large
+          trees that stress the rendering and streaming capabilities of Fizz
+        </p>
+      </section>
+      <section>
+        <h2>Another Section</h2>
+        <p>this section has a list</p>
+        <ul>
+          <li>item one</li>
+          <li>item two</li>
+          <li>item three</li>
+          <li>item four</li>
+          <li>item five</li>
+        </ul>
+        <p>it isn't a very interesting list</p>
+      </section>
+      <section>
+        <h2>Smiley Section</h2>
+        <p>here is a list of smiley emojis</p>
+        <ol>
+          <li>😀</li>
+          <li>😃</li>
+          <li>😄</li>
+          <li>😁</li>
+          <li>😆</li>
+          <li>😅</li>
+          <li>😂</li>
+          <li>🤣</li>
+          <li>🥲</li>
+          <li>☺️</li>
+          <li>😊</li>
+          <li>😇</li>
+          <li>🙂</li>
+          <li>🙃</li>
+          <li>😉</li>
+          <li>😌</li>
+          <li>😍</li>
+          <li>🥰</li>
+          <li>😘</li>
+          <li>😗</li>
+          <li>😙</li>
+          <li>😚</li>
+          <li>😋</li>
+          <li>😛</li>
+          <li>😝</li>
+          <li>😜</li>
+          <li>🤪</li>
+          <li>🤨</li>
+          <li>🧐</li>
+          <li>🤓</li>
+          <li>😎</li>
+          <li>🥸</li>
+          <li>🤩</li>
+          <li>🥳</li>
+          <li>😏</li>
+          <li>😒</li>
+          <li>😞</li>
+          <li>😔</li>
+          <li>😟</li>
+          <li>😕</li>
+          <li>🙁</li>
+          <li>☹️</li>
+          <li>😣</li>
+          <li>😖</li>
+          <li>😫</li>
+          <li>😩</li>
+          <li>🥺</li>
+          <li>😢</li>
+          <li>😭</li>
+          <li>😤</li>
+          <li>😠</li>
+          <li>😡</li>
+          <li>🤬</li>
+          <li>🤯</li>
+          <li>😳</li>
+          <li>🥵</li>
+          <li>🥶</li>
+          <li>😱</li>
+          <li>😨</li>
+          <li>😰</li>
+          <li>😥</li>
+          <li>😓</li>
+          <li>🤗</li>
+          <li>🤔</li>
+          <li>🤭</li>
+          <li>🤫</li>
+          <li>🤥</li>
+          <li>😶</li>
+          <li>😐</li>
+          <li>😑</li>
+          <li>😬</li>
+          <li>🙄</li>
+          <li>😯</li>
+          <li>😦</li>
+          <li>😧</li>
+          <li>😮</li>
+          <li>😲</li>
+          <li>🥱</li>
+          <li>😴</li>
+          <li>🤤</li>
+          <li>😪</li>
+          <li>😵</li>
+          <li>🤐</li>
+          <li>🥴</li>
+          <li>🤢</li>
+          <li>🤮</li>
+          <li>🤧</li>
+          <li>😷</li>
+          <li>🤒</li>
+          <li>🤕</li>
+          <li>🤑</li>
+          <li>🤠</li>
+          <li>😈</li>
+          <li>👿</li>
+          <li>👹</li>
+          <li>👺</li>
+          <li>🤡</li>
+          <li>💩</li>
+          <li>👻</li>
+          <li>💀</li>
+          <li>☠️</li>
+          <li>👽</li>
+          <li>👾</li>
+          <li>🤖</li>
+          <li>🎃</li>
+          <li>😺</li>
+          <li>😸</li>
+          <li>😹</li>
+          <li>😻</li>
+          <li>😼</li>
+          <li>😽</li>
+          <li>🙀</li>
+          <li>😿</li>
+          <li>😾</li>
+        </ol>
+      </section>
+      <section>
+        <h2>Translation Section</h2>
+        <p>This is the final section you will see before the sections repeat</p>
+        <p>
+          English: This is a text block translated from English to another
+          language in Google Translate.
+        </p>
+        <p>
+          Korean: 이것은 Google 번역에서 영어에서 다른 언어로 번역된 텍스트
+          블록입니다.
+        </p>
+        <p>
+          Hindi: यह Google अनुवाद में अंग्रेज़ी से दूसरी भाषा में अनुवादित
+          टेक्स्ट ब्लॉक है।
+        </p>
+        <p>
+          Lithuanian: Tai teksto blokas, išverstas iš anglų kalbos į kitą
+          „Google“ vertėjo kalbą.
+        </p>
+        <div>
+          <div>
+            <div>
+              <div>
+                <div>
+                  <div>
+                    <span>
+                      we're deep in some nested divs here, not that you can tell
+                      visually
+                    </span>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+        </div>
+      </section>
+    </article>
+  );
+}