drawing heavily from https://github.com/facebook/react/pull/24034 this PR should improve performance of renderToPipeableStream.

This PR implements a TypedArray buffer that text is encoded into while writing chunks to the stream. It also adds a fixture which is useful for comparing the baseline rendering performance of renderToPipeableStream and renderToString.

If you want to test out the current implementation simply visit the first commit.
The second commit includes an analog of the buffer which does encoding during the performWork step.
The third commit includes a change to adopt encodeInto which delivers some performance gains by avoiding extra mem copies.


### Fixture Performance
The following stats were gathered with `npx autocannon -c 5 -d 30 http://localhost:4000/{stream|string} --renderStatusCodes --excludeErrorStats`
##### Before Change
| |Avg Latency|Req per Sec|
|---:|:---:|:---:|
|String|89ms|56rps|
|Stream|2249ms|2rps|

##### buffering stream chunks
| |Avg Latency|Req per Sec|
|---:|:---:|:---:|
|String|88ms|57rps|
|Stream|250ms|20rps|

##### encodeInto
| |Avg Latency|Req per Sec|
|---:|:---:|:---:|
|String|88ms|56rps|
|Stream|146ms|34rps|