@gnoff 2022-04-07
perf stats

| | React 18 Stream | React 18 String | React 17 String |
|---:|---:|---|---|
| run 1 | pw: 114 | | |
| | fcq: 0.1 | | |
| run 2 | pw: 159 | | |
| | fcq: 0.1 | | |
| run 3 | pw: 153 | | |
| | fcq: 0.04 | | |
| run 4 | pw: 165 | | |
| | fcq: 0.04 | | |
| run 5 | pw: 123 | | |
| | fcq: 0.05 | | |