@gaearon 2022-04-08
> are docs planned for the bootstrap* options?

Yeah. I'll add it to the list for follow-ups. For now you can inspect the output HTML to see what they do.