@gnoff 2022-04-08
I think I'll delete that comment. it was performance.now timings for `performWork` and `flushCompletedQueues` to get a sense of where time was spent depending on where we did encoding. if we do it flushCompletedQueues (which is the current approach of this PR) we delay the work to when we are doing IO which per Seb may be unideal since we could do that work ahead of time. But given the overall speed improvement I think it makes sense to make this tradeoff.

If you look in the main description you will see more broad based network results from autocannon which is more real-world